feat: independant component ReuseBulb
